refactor: create useInterviewHistory custom hook
useInterviewHistory custom hook을 만들고 관련 내용을 useInterview에서 수정.
